services throw data access exceptions

+++ server/src/main/java/service/UserService.java
@@ -13,56 +13,61 @@ public class UserService {
         this.authDao = authDao;
     }
 
-    public RegisterResult register(RegisterRequest registerRequest)  { //throws AlreadyTakenException {
+    public RegisterResult register(RegisterRequest registerRequest) throws DataAccessException { //throws AlreadyTakenException {
 
-        //the problem right now is that it creates a new DAO every time it's called, so it is always empty, so there is never a user with the same username
         String username = registerRequest.username();
         String password = registerRequest.password();
         String email = registerRequest.email();
-        //right now the try works if the username can be found, so the catch is what is supposed to happen - kinda funky?
-        try {
-            userDao.getUser(username);
-//            throw new AlreadyTakenException("that username already exists!");
-            return new RegisterResult(username, "", 403);
+
+        if (userDao.getUser(username) != null) {
+            throw new DataAccessException("username already exists");
+            //403 already taken/forbidden
         }
-        catch (DataAccessException e) {
-//            System.out.println("printing exception thrown in UserService: " + e);
-            userDao.createUser(new UserData(username, password, email));
-            String authToken = authDao.createAuth(username);
-            return new RegisterResult(username, authToken, 200);
+        if (password == null || email == null) {
+            throw new DataAccessException("password or email blank");
+            //400 bad request
         }
 
+        userDao.createUser(new UserData(username, password, email));
+        String authToken = authDao.createAuth(username);
+        return new RegisterResult(username, authToken, 200);
+
     }
-    public LoginResult login(LoginRequest loginRequest) {
+
+    public LoginResult login(LoginRequest loginRequest) throws DataAccessException {
         String username = loginRequest.username();
         String password = loginRequest.password();
 
-        try {
-            UserData user = userDao.getUser(username);
-            if (password.equals(user.password())) {
-                return new LoginResult(username, authDao.createAuth(username), 200);
-            }
-            else {
-                return new LoginResult(username, "", 401);
-            }
+        if (username == null || password == null) {
+            //400 bad request
+            throw new DataAccessException("username and password are required fields.");
+        }
 
+        UserData user = userDao.getUser(username);
+
+        //credentials are correct
+        if (user != null && password.equals(user.password())) {
+            return new LoginResult(username, authDao.createAuth(username), 200);
         }
-        catch (DataAccessException e) {
-            return new LoginResult(username, "", 401);
+        else {
+            throw new DataAccessException("wrong username or password.");
+            //401 unauthorized
         }
+
     }
 
 
-    public LogoutResult logout(LogoutRequest logoutRequest) {
+    public LogoutResult logout(LogoutRequest logoutRequest) throws DataAccessException {
         String authToken = logoutRequest.authToken();
-        try {
-            authDao.getAuth(authToken);
-            authDao.deleteAuth(authToken);
-            return new LogoutResult(200);
+        //shouldn't be a problem if authToken is null? it will return 401 either way?
+
+        if (authDao.getAuth(authToken) == null) {
+            throw new DataAccessException("no auth token");
+            //401 unauthorized
         }
-        catch (DataAccessException e) {
-            return new LogoutResult(401);
-        }
+
+        authDao.deleteAuth(authToken);
+        return new LogoutResult(200);
     }
 }
